requirements.txt
python-binance
websocket-client
PyQt5
keyring